+++ inst/tinytest/test_map_p.R
@@ -27,4 +27,6 @@ mtq$POP[4:6] <- NA
 expect_message(mf_prop(x = mtq, var = "POP", add = FALSE))
 mtq$POP[8:12] <- Inf
 expect_message(mf_prop(x = mtq, var = "POP", add = FALSE))
+mtq$MED <- mtq$MED * -1
+expect_message(mf_prop(x = mtq, var = "MED", add = FALSE))